Move vcs layer to separate module
Issue #1820
