+++ weblate/requirements.py
@@ -24,10 +24,11 @@ import importlib
 import sys
 from distutils.version import LooseVersion
 from django.core.exceptions import ImproperlyConfigured
-from weblate.trans.vcs import (
-    GitRepository, HgRepository, SubversionRepository, GitWithGerritRepository,
+from weblate.vcs.git import (
+    GitRepository, SubversionRepository, GitWithGerritRepository,
     GithubRepository,
 )
+from weblate.vcs.mercurial import HgRepository
 
 
 def get_version_module(module, name, url, optional=False):
