+++ weblate/trans/models/component.py
@@ -53,7 +53,7 @@ from weblate.trans.util import (
 from weblate.trans.signals import (
     vcs_post_push, vcs_post_update, translation_post_add
 )
-from weblate.trans.vcs import RepositoryException
+from weblate.vcs.base import RepositoryException
 from weblate.vcs.models import VCS_REGISTRY
 from weblate.utils.stats import ComponentStats
 from weblate.trans.models.translation import Translation
